+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added: Browser utility `CConnoisseur.redrawBankMinigame`.
 - Added: Browser utility `CConnoisseur.startGrandmapocalypse`.
 - Added: Browser utility `CConnoisseur.spawnReindeer`.
+- Added: Browser utilities `CConnoisseur.spawnWrinkler` and `CConnoisseur.popWrinkler`.
 
 ## 0.2.0 - 2021-10-01
 - Added: Explicit support for Firefox.